FEATURE: Introduce support for link_url annotation (#11)
Previously, we had support for 'grafana_path' and 'logs_url', but these were implemented in a CDCK-specific way. This commit removes support for those, and adds a more generic system. Alerts can now include `link_url` and `link_text` annotations which will be used in the alert UI.

Three new regex site settings are introduced which can be configured to enable special handling for grafana, kibana and prometheus links. This special handling will automatically add the alert timestamps to the URLs. To disable this special handling, set the regex to an empty string.

The existing 'graph_url' column is renamed to 'generator_url' to match prometheus terminology and avoid any confusion with grafana URLs.

The existing grafana_url and logs_url columns are ignored, and will be dropped in a future commit. Data has been automatically migrated to the link_url column.