PERF: Add partial index for alert topics.
Instead of scanning all the opened topics like the previous query did,
we only scan for topics inside the configured alert categories. A
partial index is also used here so that we don't have to index other
topics unnecessarily.

On one of our own internal instance, this reduces the time spent on this
query from ~20ms to ~4ms.