DEV: Modernise JS implementation (#61)
Reimplements the alert UX to use Glimmer components and uses the RenderGlimmer helper to attach it inside the post-stream. When the post stream is eventually converted to Glimmer, this will make it much easier for this plugin to migrate.

This commit also removes the need for observers / modifyClass / appEvents. Re-rendering is now performed automatically thanks to Glimmer autotracking, and the topic controller can be injected into a Glimmer component for quote support.