UX: Ensure local-dates popover is not clipped (#66)
978e802ecb83c8397c3e0ae0d2c30173fefee381 introduced `overflow: hidden` on `.alert-table-wrapper`. That caused the local-dates popover to be clipped when it escaped the wrapper.

`position: fixed` elements like the local-dates popover are positioned based on their closest `position: relative` element. Previously, it was the `<table>` itself, which is inside the `overflow: hidden` wrapper. This commit removes that, so that the nearest `position: relative` parent is **outside** the `overflow: hidden` element, thereby avoiding the clipping.